XIVY-16411 fix outdated inscription data: spin on selection change
